move UserLayout to components
